test model on full dataset: - all tfm tested on mimic - found good params to make all run - orion really slow (unusable) - Limix slow but maybe usable (less hp to tune)

+++ src/classes/data_registry.py
@@ -32,7 +32,7 @@ class DatasetTask:
         labels = df[self.target]
         if self.target == "hours_to_readmit":
             return labels.notna().astype(int)
-        return labels
+        return labels.astype(int)
 
     def features_from(self, df: pd.DataFrame) -> pd.DataFrame:
         return df.drop(columns=list(TARGET_LIKE_COLUMNS), errors="ignore")
